pyproject.toml
[tool.poetry]
name = "defense-finder-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
biopython = "^1.79"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
